refactor: rename the files and components
