+++ application/ui/src/features/models/job-table/job-row-content.tsx
@@ -3,7 +3,7 @@ import { Heading, Item, TabList, TabPanels, Tabs, View } from '@geti-ui/ui';
 import { SchemaTrainJob } from '../../../api/openapi-spec';
 import { JobMetricsContent } from '../metrics/metrics';
 
-import classes from '../model-table/model-row-content.module.css';
+import classes from '../models-table/model-row-content.module.css';
 
 interface JobRowContentProps {
     job: SchemaTrainJob;
