+++ application/ui/src/features/models/models-table/model-row-content.module.css
