+++ application/ui/src/features/models/models-table/model-row.tsx
