+++ application/ui/src/features/models/models-table/model-table.module.css
