+++ application/ui/src/features/models/models-table/start-inference-dialog.tsx
